Refacotr according to SRP
