+++ src/main/java/com/sourcery/gymapp/backend/globalconfig/KafkaProcessingContext.java
@@ -1,0 +1,28 @@
+package com.sourcery.gymapp.backend.globalconfig;
+
+public class KafkaProcessingContext {
+    private static final ThreadLocal<Boolean> isKafkaProcessing = ThreadLocal.withInitial(() -> false);
+
+    /**
+     * Enables Kafka processing state.
+     */
+    public static void enableKafkaProcessing() {
+        isKafkaProcessing.set(true);
+    }
+
+    /**
+     * Disables Kafka processing state.
+     */
+    public static void disableKafkaProcessing() {
+        isKafkaProcessing.remove();
+    }
+
+    /**
+     * Retrieves the current Kafka processing state.
+     *
+     * @return true if Kafka processing is active, false otherwise.
+     */
+    public static boolean isKafkaProcessing() {
+        return isKafkaProcessing.get();
+    }
+}
