+++ src/main/java/com/sourcery/gymapp/backend/userProfile/consumer/UserProfileKafkaConsumer.java
@@ -3,6 +3,7 @@ package com.sourcery.gymapp.backend.userProfile.consumer;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sourcery.gymapp.backend.events.RegistrationEvent;
 import com.sourcery.gymapp.backend.globalconfig.AuditorConfig;
+import com.sourcery.gymapp.backend.globalconfig.KafkaProcessingContext;
 import com.sourcery.gymapp.backend.userProfile.service.UserProfileService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -22,7 +23,7 @@ public class UserProfileKafkaConsumer {
     @KafkaListener(topics = {"${spring.kafka.topics.account-register}"}, groupId = "user-profile-listener-group")
     public void onMessage(ConsumerRecord<UUID, String> record) {
         try {
-            AuditorConfig.AuditorAwareImpl.enableKafkaProcessing();
+            KafkaProcessingContext.enableKafkaProcessing();
 
             var data = objectMapper.readValue(record.value(), RegistrationEvent.class);
             userProfileService.createUserProfileAfterRegistration(data);
@@ -30,7 +31,7 @@ public class UserProfileKafkaConsumer {
         } catch (Exception e) {
             log.error("Error processing registration event: {}", e.getMessage(), e);
         } finally {
-            AuditorConfig.AuditorAwareImpl.disableKafkaProcessing();
+            KafkaProcessingContext.disableKafkaProcessing();
         }
     }
 }
