+++ src/main/java/com/sourcery/gymapp/backend/userProfile/service/ProfileCurrentUserService.java
@@ -3,6 +3,7 @@ package com.sourcery.gymapp.backend.userProfile.service;
 import java.util.UUID;
 
 import com.sourcery.gymapp.backend.globalconfig.AuditorConfig;
+import com.sourcery.gymapp.backend.globalconfig.KafkaProcessingContext;
 import com.sourcery.gymapp.backend.userProfile.exception.UserNotAuthenticatedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -13,7 +14,7 @@ import org.springframework.stereotype.Service;
 public class ProfileCurrentUserService {
 
     public UUID getCurrentUserId() {
-        if (AuditorConfig.AuditorAwareImpl.getKafkaProcessing()) {
+        if (KafkaProcessingContext.isKafkaProcessing()) {
             return AuditorConfig.AuditorAwareImpl.getSystemUserUUID();
         }
